Delete the old run.py to rename eager_run to run
