+++ pyroclast/prototype/README.md
@@ -1,4 +1,4 @@
 # Prototype
 
 To test:
-`python -m pyroclast.eager_run --alg prototype --dataset caltech_birds2011 --resize_data_shape 224 224 --output_dir test_prototype`
+`python -m pyroclast.run --alg prototype --dataset caltech_birds2011 --resize_data_shape 224 224 --output_dir test_prototype`
