+++ pyroclast/run.py
@@ -1,106 +1,15 @@
-import multiprocessing
+import json
+import os
 import sys
 from importlib import import_module
 
 import tensorflow as tf
+from tensorflow.python.client import device_lib
 
-import pyroclast
-from pyroclast.common.tf_util import get_session
-from pyroclast.common.data_registry import get_dataset, split_dataset, dataset_iterator
 from pyroclast.common.cmd_util import common_arg_parser, parse_unknown_args
+from pyroclast.common.tf_util import setup_tfds
 
-
-def get_task_type(task_id):
-    tasks = pyroclast.common.data_registry.tasks
-
-    task_type = None
-    for t, ids in tasks.items():
-        if task_id in ids:
-            task_type = t
-            break
-    assert task_type is not None, 'task_id {} is not recognized in task types'.format(
-        task_id)
-
-    return task_type, task_id
-
-
-def get_alg_module(alg, submodule=None):
-    submodule = submodule or alg
-    try:
-        # first try to import the alg module from baselines
-        alg_module = import_module('.'.join(['pyroclast',
-                                             alg]))  #, submodule]))
-    except ImportError:
-        # then from rl_algs
-        # alg_module = import_module('.'.join(['rl_' + 'algs', alg, submodule]))
-        print('failed to import from {}'.format('.'.join(['pyroclast', alg])))
-        assert False
-
-    return alg_module
-
-
-def get_learn_function(alg):
-    return get_alg_module(alg).learn
-
-
-def get_learn_function_defaults(alg, env_type):
-    try:
-        alg_defaults = get_alg_module(alg, 'defaults')
-        kwargs = getattr(alg_defaults, env_type)()
-    except (ImportError, AttributeError):
-        kwargs = {}
-    return kwargs
-
-
-def setup_data(args):
-    dataset_fn = get_dataset(args.dataset)
-    ds_t = lambda ds: dataset_iterator(ds, args.epochs, args.batch_size)
-    if args.task == 'semisupervised':
-        assert args.num_labeled is not None
-        train_data = dataset_fn()['train']
-        labeled, unlabeled = split_dataset(train_data, args.num_labeled)
-        labeled = ds_t(labeled)
-        unlabeled = ds_t(unlabeled)
-        train_data = (labeled, unlabeled)
-    else:
-        dataset_dict, info = dataset_fn()
-        train_data_iterator = ds_t(dataset_dict['train'])
-        train_batches_per_epoch = info.splits[
-            'train'].num_examples // args.batch_size
-        test_data_iterator = ds_t(dataset_dict['test'])
-        test_batches_per_epoch = info.splits[
-            'test'].num_examples // args.batch_size
-    return train_data_iterator, train_batches_per_epoch, test_data_iterator, test_batches_per_epoch
-
-
-def train(args, extra_args):
-    # load data
-    task_type, data_name = get_task_type(args.dataset)
-    print('task_type: {}'.format(task_type))
-
-    seed = args.seed
-    train_data_iterator, train_batches_per_epoch, test_data_iterator, test_batches_per_epoch = setup_data(
-        args)
-
-    learn = get_learn_function(args.alg)
-    alg_kwargs = get_learn_function_defaults(args.alg, task_type)
-    alg_kwargs.update(extra_args)
-
-    print('Training {} on {} with arguments \n{}'.format(
-        args.alg, data_name, alg_kwargs))
-
-    model = learn(train_data_iterator=train_data_iterator,
-                  train_batches_per_epoch=train_batches_per_epoch,
-                  test_data_iterator=test_data_iterator,
-                  test_batches_per_epoch=test_batches_per_epoch,
-                  seed=seed,
-                  **alg_kwargs)
-
-    return model
-
-
-def get_default_network(task_type):
-    return 'cnn'
+tf.compat.v1.enable_eager_execution()
 
 
 def parse_cmdline_kwargs(args):
@@ -120,10 +29,69 @@ def parse_cmdline_kwargs(args):
     return {k: parse(v) for k, v in parse_unknown_args(args).items()}
 
 
+def get_alg_module(alg, submodule=None):
+    submodule = submodule or alg
+    try:
+        # first try to import the alg module from baselines
+        alg_module = import_module('.'.join(['pyroclast', alg, submodule]))
+    except ImportError:
+        print('failed to import from {}'.format('.'.join(['pyroclast', alg])))
+        assert False
+
+    return alg_module
+
+
+def get_learn_function(alg):
+    return get_alg_module(alg).learn
+
+
+def get_learn_function_defaults(alg, dataset):
+    try:
+        alg_defaults = get_alg_module(alg, 'defaults')
+        kwargs = getattr(alg_defaults, dataset)()
+    except (ImportError, AttributeError):
+        kwargs = {}
+    return kwargs
+
+
+def train(args, extra_args):
+    # load data
+    learn = get_learn_function(args.alg)
+    alg_kwargs = get_learn_function_defaults(args.alg, args.dataset)
+    alg_kwargs.update(extra_args)
+    data_dict = setup_tfds(args.dataset, args.batch_size,
+                           args.resize_data_shape, args.data_limit,
+                           args.data_dir)
+
+    print('Training {} on {} with arguments \n{}'.format(
+        args.alg, args.dataset, alg_kwargs))
+    model = learn(data_dict,
+                  seed=args.seed,
+                  output_dir=args.output_dir,
+                  debug=args.debug,
+                  **alg_kwargs)
+
+    return model
+
+
 def main(args):
     arg_parser = common_arg_parser()
     args, unknown_args = arg_parser.parse_known_args(args)
     extra_args = parse_cmdline_kwargs(unknown_args)
+
+    if args.debug:
+        print(device_lib.list_local_devices())
+
+    # save the parameters
+    if not os.path.exists(args.output_dir):
+        os.makedirs(args.output_dir)
+    with open(os.path.join(args.output_dir, 'parameters.log'), 'w') as p_file:
+        json.dump(
+            {
+                'args': args.__dict__,
+                'unknown_args': unknown_args,
+                'extra_args': extra_args
+            }, p_file)
 
     model = train(args, extra_args)
